Update annotations
Fixed some annotation mistakes in the rewrite_examples output for example_0001.xml. Note that for object_property and abstract_container symbols, the expression of all the symbols' properties as individual symbols is typically missing something. Also, the container_type was incorrect for some of the container symbols and a container_type symbol was incorrectly expressed as an empty container.