Update example_0013.xml annotations
Changed an empty container to a container_type symbol. Changed an abstract_container to a region_abstract_container symbol.

Modified the parse tree at the phrase "from the robot". An analogous navigation dataset example: objects_in_groups/example_0024.xml

"Robot" grounds to a container in this corpus because we represent the robot as both of the hands; I think the object_type="robot" is important for learning this difference between the corpora. This representation is useful in anticipation of commands like "use your right arm to pick up the block in the middle".