Update example_0015.xml annotations
This parse tree representation may need to change; "on the table" seems to modify "the column" rather than "four red blocks". Also, we may want to consider "on the table" aa region with a spatial_relation_type "on". Finally, depending on how we address these points, we may want to change which "index" and "number" symbols get carried up the tree.