Update example_0018.xml annotations
Commented the abstract container symbol - "the red blocks" is now expressed as a container with each red block from the world model as a member object.

We may want to reconsider the parse tree structure at "on the table". Also, we want to express "on the table" as a region.