Update example_0040 - example_0049 annotations
Most examples had an abstract container and related symbols expressed at a leaf phrase like "five blocks" -in the context of the world model, the specific objects are immediately resolvable and so are expresed as a container, the objects, and the object type.

example_0047 - example_0049 have parse tree structures that we may want to reconsider. For example, example_0047 treats "in front of..." differently than example_0046.