Update annotations for example_0060 - example_0069
Reverted a container type to "group" in example_0017.

Changed the parse tree structure for each example at phrases like "from the right" to make them siblings of "in front of...". This is consistent with the navigation dataset.

Example_0066 expresses two spatial_relation symbols at the phrase "the cone furthest right". Only one spatial_relation is expressed in the coresponding object_property symbol (right, in this case) - this is consistent with the navigation dataset. Analogous navigation example: objects_in_groups/example_0012.xml