Update matlab accuracy plot script generation
In cases where the results are not ordered according to the cross validation partitions sizes, the plots can have spurious trend lines. So, there is now additional matlab code to sort the tests according to the partition size before plotting.

Additionally, the plots of average root && average complete accuracies did not include the error bars, so the 'plot' command has been replaced with an equivalent 'errorbar' command. This is relevant for figures 9 - 12.